feat: Refactor envío de recibo y extraer lógica a funciones utilitarias
