+++ src/interfaces/Filtro.ts
