+++ src/utils/sendEmails.ts
@@ -1,0 +1,14 @@
+import { bodyRecibo } from "../emails/recibido";
+import { ReparacionType } from "../types/reparacion";
+import { enviarEmail } from "./utils";
+
+export const enviarRecibo = (reparacion: ReparacionType): void => {
+    const datosEmail = {
+        to: reparacion.data.EmailUsu,
+        cc: 'info@mauriciocruzdrones.com',
+        bcc: [],
+        subject: 'Recibo de equipo ' + reparacion.data.DroneRep,
+        body: bodyRecibo(reparacion),
+    };
+    enviarEmail(datosEmail);
+}
